# Improve performance of `--source-dir`