@marco-c 2025-04-03
> Still, I hope the speedup so far is appreciated :D

Very appreciated!
Unfortunately the rewrite paths stuff is a bit overly complex and brittle.