@marco-c 2025-04-04
Filed #1308 as a potential follow-up optimization.